Refactor event loop to use Agent object rather than individual parameters (#359)
Update the event_loop_cycle function to accept the agent directly instead of the separate parameters that were coming directly from the agent. This simplifies + clarifies exactly what the event loop is doing with the agent.  Long-term this sets us up for more easily access other agent capabilities like hooks.

There is a downside here of tighter coupling between the agent & the event_loop - an alternative implementation of this would be to abstract the object being passed - maybe something like `EventLoopContext`.  I originally started implementing this approach before revisiting it as it seemed like an unnecessary abstraction.